Extract Get extension from filename tests
